examples/bedrock-agent/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
